Add test set during training.
We now reserve some samples from the "train" set so that we can observe
generalization performance during training. We can't do this with the
"test" or "validation" sets of the meta-learning datasets, because they
never share the same classes with the training set. We want to see how
well the model generalizes to never-seen instances from seen classes.