Change to 1-based steps.
This numbering scheme makes more sense in general, but we didn't use it
before for some arcane reasons. Now we have an option
--save-initial-model that serves that purpose.

Now that we have 1-based steps, we can evaluate performance at the _end_
of each epoch, rather than the _beginning_. This is important especially
when applying lots of lobotomy, because before we would have saved and
evaluated the model just after lobotomy, making it an impaired
evaluation.